requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
urllib3==2.2.3
